backend/requirements.txt
# backend/requirements.txt

# Core
Django==5.0.6
djangorestframework==3.15.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
django-celery-beat==2.6.0
django-celery-results==2.5.1
celery==5.4.0
redis==5.0.1
Pillow
# Geração de PDF
weasyprint